refactor(4713): decompose voice barge-in giant
