+++ ARCHITECTURE.md
@@ -920,8 +920,12 @@ src/
 │   │   │   ├── progress_playback.rs
 │   │   │   ├── receive_hook.rs
 │   │   │   ├── routing.rs
+│   │   │   ├── runtime_lifecycle.rs
 │   │   │   ├── stt.rs
-│   │   │   └── tts_pipeline.rs
+│   │   │   ├── tts_pipeline.rs
+│   │   │   ├── turn_dispatch.rs
+│   │   │   ├── utility.rs
+│   │   │   └── utterance_pipeline.rs
 │   │   ├── watchers/
 │   │   │   ├── lifecycle/
 │   │   │   │   └── activity.rs
