+++ src/services/discord/voice_barge_in.rs
@@ -43,7 +43,12 @@ use super::voice_background_driver::{
 use super::voice_config_cache::ConfigSnapshotCache;
 use super::voice_id_sequences::VoiceIdSequences;
 use super::voice_sensitivity::SensitivityState;
-use super::{SharedData, http, mailbox_has_active_turn, rate_limit_wait, settings};
+pub(super) use super::{
+    SharedData, enqueue_internal_followup, http,
+    mailbox_cancel_active_turn_if_handoff_user_message_with_reason,
+    mailbox_cancel_active_turn_with_reason, mailbox_has_active_turn, rate_limit_wait,
+    record_voice_handoff_cancel_tombstone, settings,
+};
 
 #[path = "voice_barge_in/channel_state.rs"]
 mod channel_state;
@@ -72,26 +77,26 @@ mod routing;
 use routing::{
     agent_voice_background_channel, agent_voice_matches_channel, effective_voice_source_channel,
 };
+#[path = "voice_barge_in/runtime_lifecycle.rs"]
+mod runtime_lifecycle;
 #[path = "voice_barge_in/stt.rs"]
 mod stt;
 #[path = "voice_barge_in/tts_pipeline.rs"]
 mod tts_pipeline;
+#[path = "voice_barge_in/turn_dispatch.rs"]
+mod turn_dispatch;
+#[path = "voice_barge_in/utility.rs"]
+mod utility;
+#[path = "voice_barge_in/utterance_pipeline.rs"]
+mod utterance_pipeline;
+pub(in crate::services::discord) use utility::{
+    INTERNAL_VOICE_MESSAGE_ID_START, is_synthetic_voice_message_id,
+};
+use utility::{
+    discord_pcm_i16_stereo_48k_to_mono_f32_16k, lock_monitor, pcm_i16_to_le_bytes,
+    remove_file_quietly, remove_file_quietly_silent, transcript_dirs_from_config,
+};
 
-pub(in crate::services::discord) const INTERNAL_VOICE_MESSAGE_ID_START: u64 =
-    9_000_000_000_000_000_000;
-
-/// `true` iff `msg_id` is a synthetic voice-originated id (≥
-/// `INTERNAL_VOICE_MESSAGE_ID_START`). Real Discord snowflakes encode
-/// timestamps and worker/process/sequence fields and stay well below 2^63
-/// for the foreseeable future, so the 9e18 prefix is safely above them.
-/// Used by the message intake to skip ⏳/📬 reactions, placeholder POSTs,
-/// and `message_reference` lookups that would fail with "Unknown message"
-/// for a non-existent Discord message id.
-pub(in crate::services::discord) fn is_synthetic_voice_message_id(
-    msg_id: poise::serenity_prelude::MessageId,
-) -> bool {
-    msg_id.get() >= INTERNAL_VOICE_MESSAGE_ID_START
-}
 const STT_TRANSCRIPT_POLL_TIMEOUT: Duration = Duration::from_secs(5);
 const STT_TRANSCRIPT_POLL_INTERVAL: Duration = Duration::from_millis(200);
 const PROCESSING_CHIME_FILE_NAME: &str = "agentdesk-voice-processing-chime.wav";
@@ -992,1888 +997,6 @@ impl Drop for InflightForegroundCancelGuard<'_> {
         self.runtime
             .unregister_inflight_foreground_cancel(self.channel_id, &self.token);
     }
-}
-
-impl VoiceBargeInRuntime {
-    pub(in crate::services::discord) fn from_voice_config(config: &VoiceConfig) -> Self {
-        let default_sensitivity = config.barge_in.sensitivity;
-        let conservative_ttl = Duration::from_secs(config.barge_in.conservative_ttl_secs.max(1));
-        let stt = if config.enabled {
-            Some(VoiceSttRuntime::from_voice_config(config))
-        } else {
-            None
-        };
-        let tts = if config.enabled {
-            TtsRuntime::from_voice_config(config).ok()
-        } else {
-            None
-        };
-        let (progress_tx, _) = broadcast::channel(128);
-
-        Self {
-            enabled: config.enabled,
-            barge_in_enabled: config.enabled && config.barge_in.enabled,
-            sensitivity: SensitivityState::new(default_sensitivity, conservative_ttl),
-            acknowledgement: AcknowledgementConfig::from_voice_config(config),
-            transcript_dirs: transcript_dirs_from_config(config),
-            voice_config_state: RwLock::new(config.clone()),
-            spoken_result_language: RwLock::new(config.stt.language.clone()),
-            verbose_progress: AtomicBool::new(config.verbose_progress),
-            streaming_stt_enabled: AtomicBool::new(
-                stt.as_ref().is_some_and(VoiceSttRuntime::is_streaming),
-            ),
-            stt: RwLock::new(stt),
-            streaming_stt: StreamingSttSessions::new(),
-            tts: RwLock::new(tts),
-            progress_tx,
-            channels: VoiceChannelStateMachines::new(),
-            id_sequences: VoiceIdSequences::new(),
-            config_cache: ConfigSnapshotCache::new(),
-            alias_collision_signature: std::sync::Mutex::new(None),
-            #[cfg(test)]
-            test_state: Arc::new(VoiceBargeInTestState::default()),
-        }
-    }
-
-    #[allow(dead_code)] // #3034: test-only runtime constructor; no production caller
-    pub(in crate::services::discord) fn disabled() -> Self {
-        let (progress_tx, _) = broadcast::channel(128);
-        Self {
-            enabled: false,
-            barge_in_enabled: false,
-            sensitivity: SensitivityState::disabled(),
-            acknowledgement: AcknowledgementConfig::disabled(),
-            transcript_dirs: Vec::new(),
-            voice_config_state: RwLock::new(VoiceConfig::default()),
-            spoken_result_language: RwLock::new(DEFAULT_STT_LANGUAGE.to_string()),
-            verbose_progress: AtomicBool::new(false),
-            stt: RwLock::new(None),
-            streaming_stt: StreamingSttSessions::new(),
-            streaming_stt_enabled: AtomicBool::new(false),
-            tts: RwLock::new(None),
-            progress_tx,
-            channels: VoiceChannelStateMachines::new(),
-            id_sequences: VoiceIdSequences::new(),
-            config_cache: ConfigSnapshotCache::new(),
-            alias_collision_signature: std::sync::Mutex::new(None),
-            #[cfg(test)]
-            test_state: Arc::new(VoiceBargeInTestState::default()),
-        }
-    }
-
-    async fn generate_foreground_ack_text_for_runtime(
-        &self,
-        transcript: &str,
-        language: &str,
-        foreground: &EffectiveVoiceForegroundConfig,
-        cancel_token: Arc<crate::services::provider::CancelToken>,
-    ) -> Option<VoiceForegroundDecision> {
-        #[cfg(test)]
-        if let Some(decision) = self
-            .test_state
-            .foreground_decisions
-            .lock()
-            .expect("voice test foreground decisions lock")
-            .pop_front()
-        {
-            return Some(decision);
-        }
-
-        generate_foreground_ack_text(transcript, language, foreground, cancel_token).await
-    }
-
-    async fn generate_voice_background_result_summary_for_runtime(
-        &self,
-        background_result: &str,
-        language: &str,
-        foreground: &EffectiveVoiceForegroundConfig,
-        cancel_token: Arc<crate::services::provider::CancelToken>,
-    ) -> Option<String> {
-        #[cfg(test)]
-        if let Some(summary) = self
-            .test_state
-            .background_result_summaries
-            .lock()
-            .expect("voice test background summaries lock")
-            .pop_front()
-        {
-            return summary;
-        }
-
-        generate_voice_background_result_summary(
-            background_result,
-            language,
-            foreground,
-            cancel_token,
-        )
-        .await
-    }
-
-    #[cfg(test)]
-    fn take_test_background_handoff_outcome(
-        &self,
-        driver_kind: VoiceBackgroundDriverKind,
-        source_channel_id: ChannelId,
-        target_channel_id: ChannelId,
-        announcement: &crate::voice::prompt::VoiceTranscriptAnnouncement,
-        summary: &str,
-        message_content: &str,
-    ) -> Option<Result<VoiceBackgroundStartOutcome, String>> {
-        let outcome = self
-            .test_state
-            .background_handoff_outcomes
-            .lock()
-            .expect("voice test background handoff outcomes lock")
-            .pop_front()?;
-        self.test_state
-            .background_starts
-            .lock()
-            .expect("voice test background starts lock")
-            .push(TestVoiceBackgroundStart {
-                driver_kind,
-                source_channel_id,
-                target_channel_id,
-                utterance_id: announcement.utterance_id.clone(),
-                summary: summary.to_string(),
-                message_content: message_content.to_string(),
-            });
-        Some(outcome)
-    }
-
-    #[cfg(test)]
-    fn take_test_turn_start_outcome(
-        &self,
-        driver_kind: VoiceBackgroundDriverKind,
-        source_channel_id: ChannelId,
-        target_channel_id: ChannelId,
-        utterance_id: &str,
-        message_content: &str,
-    ) -> Option<Result<VoiceBackgroundStartOutcome, String>> {
-        let outcome = self
-            .test_state
-            .turn_start_outcomes
-            .lock()
-            .expect("voice test turn start outcomes lock")
-            .pop_front()?;
-        self.test_state
-            .turn_starts
-            .lock()
-            .expect("voice test turn starts lock")
-            .push(TestVoiceBackgroundStart {
-                driver_kind,
-                source_channel_id,
-                target_channel_id,
-                utterance_id: utterance_id.to_string(),
-                summary: String::new(),
-                message_content: message_content.to_string(),
-            });
-        Some(outcome)
-    }
-
-    pub(in crate::services::discord) fn enabled(&self) -> bool {
-        self.enabled
-    }
-
-    #[allow(dead_code)] // #3034: test-only mutable-state inspector; no production caller
-    pub(in crate::services::discord) async fn runtime_config_snapshot(
-        &self,
-    ) -> VoiceRuntimeConfigSnapshot {
-        let config = self.voice_config_state.read().await;
-        let mut snapshot = VoiceRuntimeConfigSnapshot::from(&*config);
-        snapshot.verbose_progress = self.verbose_progress_enabled();
-        snapshot
-    }
-
-    /// #2250: register a CancelToken for an in-flight foreground/voice Codex
-    /// or Claude call so explicit-stop barge-in, supersession by a new
-    /// utterance, or runtime cleanup can terminate the spawned child.
-    fn register_inflight_foreground_cancel(
-        &self,
-        channel_id: ChannelId,
-        token: Arc<crate::services::provider::CancelToken>,
-    ) {
-        self.channels
-            .inflight_foreground_cancels
-            .entry(channel_id.get())
-            .or_default()
-            .push(token);
-    }
-
-    /// #2250: remove a previously registered CancelToken once the
-    /// foreground call has returned (cancelled or completed normally).
-    fn unregister_inflight_foreground_cancel(
-        &self,
-        channel_id: ChannelId,
-        token: &Arc<crate::services::provider::CancelToken>,
-    ) {
-        if let Some(mut entry) = self
-            .channels
-            .inflight_foreground_cancels
-            .get_mut(&channel_id.get())
-        {
-            entry.retain(|existing| !Arc::ptr_eq(existing, token));
-        }
-        self.channels
-            .inflight_foreground_cancels
-            .remove_if(&channel_id.get(), |_, value| value.is_empty());
-    }
-
-    /// #3910: whether the live STT runtime is in streaming mode. Read
-    /// synchronously from an atomic mirror so the per-PCM-tick hook can gate
-    /// streaming work without awaiting the `stt` lock. Kept in sync wherever
-    /// `stt` is (re)built.
-    pub(in crate::services::discord) fn streaming_stt_enabled(&self) -> bool {
-        self.streaming_stt_enabled.load(Ordering::Relaxed)
-    }
-
-    /// #2250: signal cancellation on every in-flight foreground call for the
-    /// given channel. Called by explicit-stop barge-in and supersession
-    /// paths so the spawned Codex/Claude child is killed instead of running
-    /// to natural exit (ADR #2175).
-    pub(in crate::services::discord) fn cancel_inflight_foreground_calls(
-        &self,
-        channel_id: ChannelId,
-        reason: &'static str,
-    ) -> usize {
-        let Some((_, tokens)) = self
-            .channels
-            .inflight_foreground_cancels
-            .remove(&channel_id.get())
-        else {
-            return 0;
-        };
-        let count = tokens.len();
-        for token in tokens {
-            token.publish_cancel(reason);
-            token.cancel_with_tmux_cleanup();
-        }
-        if count > 0 {
-            tracing::info!(
-                channel_id = channel_id.get(),
-                count,
-                reason,
-                "voice foreground inflight Codex/Claude calls cancelled (#2250)"
-            );
-        }
-        count
-    }
-
-    pub(in crate::services::discord) fn verbose_progress_enabled(&self) -> bool {
-        self.verbose_progress.load(Ordering::Relaxed)
-    }
-
-    pub(in crate::services::discord) fn set_verbose_progress_enabled(&self, enabled: bool) {
-        self.verbose_progress.store(enabled, Ordering::Relaxed);
-    }
-
-    async fn spoken_result_language(&self) -> String {
-        self.spoken_result_language.read().await.clone()
-    }
-
-    pub(in crate::services::discord) async fn try_handle_voice_channel_text_reply(
-        &self,
-        http: &Arc<serenity::http::Http>,
-        channel_id: ChannelId,
-        text: &str,
-    ) -> bool {
-        let text = text.trim();
-        if text.is_empty() {
-            return false;
-        }
-
-        let config = self.cached_config().await;
-        let Some(target_channel_id) = config.agents.iter().find_map(|agent| {
-            agent_voice_matches_channel(agent, channel_id)
-                .then(|| agent_voice_background_channel(agent).unwrap_or(channel_id))
-        }) else {
-            return false;
-        };
-        drop(config);
-
-        let language = self.spoken_result_language().await;
-        let foreground = self
-            .resolve_effective_foreground_config(channel_id, target_channel_id)
-            .await;
-        let cancel_token = Arc::new(crate::services::provider::CancelToken::new());
-        // #3911 + #2335 (c): the guard OWNS the registration. It registers the
-        // token BEFORE the generate `.await` below and unregisters on drop, so
-        // an abort mid-`.await` (shutdown / supervisor abort) still runs the
-        // Drop and cannot leak the token in `inflight_foreground_cancels` (a
-        // leak would keep `has_inflight_foreground` true and misclassify the
-        // next fresh utterance as a barge-in). Keeping the guard alive through
-        // the `channel_id.say` HTTP call below still lets a late cancel
-        // suppress the now-stale reply.
-        let _text_reply_guard =
-            InflightForegroundCancelGuard::register(self, channel_id, cancel_token.clone());
-        let reply =
-            generate_voice_channel_text_reply(text, &language, &foreground, cancel_token.clone())
-                .await
-                .unwrap_or_else(|| {
-                    "지금 보이스 빠른 답변 모델 응답을 만들지 못했어요.".to_string()
-                });
-
-        if cancel_token.cancelled.load(Ordering::Relaxed) {
-            tracing::info!(
-                channel_id = channel_id.get(),
-                cancel_source = ?cancel_token.cancel_source(),
-                cancel_source_kind = ?cancel_token.cancel_source_kind(),
-                stage = "pre_post",
-                "voice channel text reply suppressed because cancel won the race (#2335)"
-            );
-            return true;
-        }
-
-        if let Err(error) = channel_id.say(http.as_ref(), reply).await {
-            tracing::warn!(
-                error = %error,
-                channel_id = channel_id.get(),
-                foreground_provider = %foreground.provider,
-                foreground_model = %foreground.model,
-                "failed to send voice channel text reply"
-            );
-        }
-        true
-    }
-
-    async fn runtime_wake_word_decision(&self, transcript: &str) -> WakeWordDecision {
-        let config = self.voice_config_state.read().await;
-        wake_word_decision(transcript, &config.wake_words, config.wake_word_required())
-    }
-
-    async fn apply_dispatcher_command(
-        &self,
-        channel_id: ChannelId,
-        transcript: &str,
-    ) -> Option<VoiceBargeInTranscriptOutcome> {
-        match parse_voice_command(transcript)? {
-            VoiceCommand::Sensitivity(sensitivity) => {
-                self.set_sensitivity(sensitivity).await;
-                tracing::info!(
-                    channel_id = channel_id.get(),
-                    sensitivity = ?sensitivity,
-                    "voice barge-in sensitivity changed by spoken command"
-                );
-                Some(VoiceBargeInTranscriptOutcome::SensitivityChanged(
-                    sensitivity,
-                ))
-            }
-            VoiceCommand::VerboseProgress(enabled) => {
-                self.set_verbose_progress_enabled(enabled);
-                tracing::info!(
-                    channel_id = channel_id.get(),
-                    verbose_progress = enabled,
-                    "voice verbose progress changed by spoken command"
-                );
-                Some(VoiceBargeInTranscriptOutcome::VerboseProgressChanged { enabled })
-            }
-            VoiceCommand::Language(language) => {
-                self.set_runtime_language(language.clone()).await;
-                Some(VoiceBargeInTranscriptOutcome::LanguageChanged(language))
-            }
-            VoiceCommand::TtsVoice(voice) => {
-                self.set_runtime_tts_voice(voice.clone()).await;
-                Some(VoiceBargeInTranscriptOutcome::TtsVoiceChanged(voice))
-            }
-            VoiceCommand::VoiceClone { reference } => {
-                tracing::info!(
-                    channel_id = channel_id.get(),
-                    reference = ?reference,
-                    "voice clone command accepted for downstream implementation"
-                );
-                Some(VoiceBargeInTranscriptOutcome::VoiceCloneRequested { reference })
-            }
-            VoiceCommand::WakeWords(command) => {
-                let wake_words = self.apply_wake_word_command(command).await;
-                let required = self.voice_config_state.read().await.wake_word_required();
-                Some(VoiceBargeInTranscriptOutcome::WakeWordsChanged {
-                    required,
-                    wake_words,
-                })
-            }
-        }
-    }
-
-    // F8 (#2046): 텍스트 디스패처(`!vc <subcommand>`)도 음성 디스패처와 동일하게
-    // Language/TtsVoice/VoiceClone/WakeWords 명령을 모두 적용할 수 있도록 setter
-    // 들을 노출한다.
-    pub(in crate::services::discord) async fn set_runtime_language_external(
-        &self,
-        language: String,
-    ) {
-        self.set_runtime_language(language).await;
-    }
-
-    pub(in crate::services::discord) async fn set_runtime_tts_voice_external(&self, voice: String) {
-        self.set_runtime_tts_voice(voice).await;
-    }
-
-    pub(in crate::services::discord) async fn apply_wake_word_command_external(
-        &self,
-        command: WakeWordCommand,
-    ) -> Vec<String> {
-        self.apply_wake_word_command(command).await
-    }
-
-    async fn set_runtime_language(&self, language: String) {
-        let config = {
-            let mut config = self.voice_config_state.write().await;
-            config.stt.language = language.clone();
-            config.clone()
-        };
-        *self.spoken_result_language.write().await = language;
-        if self.enabled {
-            self.streaming_stt.clear();
-            let runtime = VoiceSttRuntime::from_voice_config(&config);
-            // #3910: keep the synchronous streaming mirror aligned with the
-            // freshly rebuilt runtime before publishing it.
-            self.streaming_stt_enabled
-                .store(runtime.is_streaming(), Ordering::Relaxed);
-            *self.stt.write().await = Some(runtime);
-        }
-    }
-
-    async fn apply_wake_word_command(&self, command: WakeWordCommand) -> Vec<String> {
-        let mut config = self.voice_config_state.write().await;
-        match command {
-            WakeWordCommand::EnableDefault => {
-                if config
-                    .wake_words
-                    .iter()
-                    .all(|value| value.trim().is_empty())
-                {
-                    config.wake_words = vec![DEFAULT_WAKE_WORD.to_string()];
-                }
-            }
-            WakeWordCommand::Disable => {
-                config.wake_words.clear();
-            }
-            WakeWordCommand::Set(wake_words) => {
-                config.wake_words = wake_words;
-            }
-        }
-        config.wake_words.clone()
-    }
-
-    pub(in crate::services::discord) async fn unregister_voice_guild(&self, guild_id: GuildId) {
-        // F7 (#2046): voice_guilds 만 지우면 channel_id 키로 적재된 monitors /
-        // playbacks / spoken_result_playbacks / active_voice_routes /
-        // deferred_buffers 가 남아 join/leave 반복 시 누수. 같은 guild 의 모든
-        // control_channel_id 를 먼저 수집해 채널 단위 state 도 함께 정리한다.
-        let stale_channels = self.channels.remove_guild_contexts(guild_id);
-        // #3910: outer handles for the leaving channels, collected so the inner
-        // WhisperStream sessions can be discarded after the sync teardown loop.
-        let mut stranded_stream_sessions: Vec<SttSessionHandle> = Vec::new();
-        for channel_id in stale_channels {
-            self.channels.disconnected(ChannelId::new(channel_id));
-            self.channels.monitors.remove(&channel_id);
-            if let Some((_, session)) = self.channels.playbacks.remove(&channel_id) {
-                session.cancellation.cancel();
-            }
-            if let Some((_, session)) = self.channels.spoken_result_playbacks.remove(&channel_id) {
-                session.cancellation.cancel();
-            }
-            // #2250: also abort any in-flight foreground Codex/Claude call so
-            // its spawned child does not outlive the guild teardown.
-            self.cancel_inflight_foreground_calls(
-                ChannelId::new(channel_id),
-                "voice_guild_teardown",
-            );
-            self.channels.active_voice_routes.remove(&channel_id);
-            self.channels.deferred_buffers.remove(&channel_id);
-            // #3910: a speaker leaving the channel mid-utterance otherwise leaves
-            // its streaming-STT session + feed-task bucket stranded in the maps
-            // (they were only reaped at utterance completion). Drop the outer
-            // bucket + abort pending feed tasks here, and collect the outer
-            // handles so the inner stream session is reaped below too.
-            stranded_stream_sessions.extend(self.streaming_stt.remove_channel(channel_id));
-            self.channels.forget(channel_id);
-        }
-        // #3910: dropping the outer `SttSessionHandle` alone leaves the matching
-        // inner `WhisperStream` session (inserted by `start_session`, removed
-        // only by `finalize()`) leaked until the runtime is rebuilt. Discard
-        // those inner sessions for the leaving channels (no final decode — the
-        // speaker left, so partial-transcript loss is acceptable). File mode
-        // keeps no inner session, so this is a no-op there.
-        if !stranded_stream_sessions.is_empty() {
-            // Hoist read+clone into a local so the `RwLockReadGuard` drops at
-            // this statement's end — NOT held across the `discard_stream_session`
-            // awaits below (an `if let` scrutinee would extend the guard over the
-            // whole body, holding `self.stt` read while awaiting → deadlock risk
-            // against any `self.stt` writer / re-entrant lock). A plain `let`
-            // binding (not `let-else` + `return`) preserves the rest of the fn.
-            let stt = self.stt.read().await.clone();
-            if let Some(stt) = stt {
-                for handle in stranded_stream_sessions {
-                    stt.discard_stream_session(&handle).await;
-                }
-            }
-        }
-    }
-
-    /// F2 (#2046): 특정 길드에 매핑된 control_channel_id 목록을 반환.
-    /// `leave_voice_channel` 경로에서 `VoiceReceiver::flush_for_control_channel`을
-    /// 길드 단위로 한정 호출하기 위해 사용한다.
-    pub(in crate::services::discord) fn control_channel_ids_for_guild(
-        &self,
-        guild_id: GuildId,
-    ) -> Vec<u64> {
-        self.channels.channel_ids_for_guild(guild_id)
-    }
-
-    pub(in crate::services::discord) fn spawn_sensitivity_ttl_reset(
-        self: &Arc<Self>,
-        shutdown_flag: Arc<AtomicBool>,
-    ) {
-        if !self.barge_in_enabled {
-            return;
-        }
-
-        let state = self.sensitivity.state_handle();
-        let token = CancellationToken::new();
-        let reset_token = token.clone();
-        tokio::spawn(run_sensitivity_ttl_reset(state, reset_token));
-        // F21 (#2046): shutdown_flag 폴링 주기를 1초 → 5초 로 늘려 cpu wakeup 비용을
-        // 1/5 로 줄인다. shutdown 전체 latency 가 최대 5초 늘지만 sensitivity TTL
-        // 자체가 분 단위 주기라 실효 영향이 거의 없다. (Full CancellationToken
-        // 통합은 SharedData 차원 리팩토링이라 follow-up 으로 남긴다.)
-        tokio::spawn(async move {
-            while !shutdown_flag.load(Ordering::Relaxed) {
-                tokio::time::sleep(Duration::from_secs(5)).await;
-            }
-            token.cancel();
-        });
-    }
-
-    pub(in crate::services::discord) async fn set_sensitivity(
-        &self,
-        sensitivity: BargeInSensitivity,
-    ) {
-        self.sensitivity.set(sensitivity).await;
-        self.update_existing_monitor_sensitivity(sensitivity);
-    }
-
-    #[allow(dead_code)] // #3034: test-only voice-command entry point; no production caller
-    pub(in crate::services::discord) async fn apply_voice_command(
-        &self,
-        transcript: &str,
-    ) -> Option<BargeInSensitivity> {
-        if !self.barge_in_enabled {
-            return None;
-        }
-        let sensitivity = self.sensitivity.apply_voice_command(transcript).await?;
-        self.update_existing_monitor_sensitivity(sensitivity);
-        Some(sensitivity)
-    }
-
-    pub(in crate::services::discord) async fn handle_processing_transcript(
-        &self,
-        shared: &Arc<SharedData>,
-        _provider: &ProviderKind,
-        channel_id: ChannelId,
-        transcript: &str,
-    ) -> VoiceBargeInTranscriptOutcome {
-        if !self.enabled {
-            return VoiceBargeInTranscriptOutcome::Disabled;
-        }
-
-        let transcript = transcript.trim();
-        if transcript.is_empty() {
-            return VoiceBargeInTranscriptOutcome::EmptyTranscript;
-        }
-
-        if !self.barge_in_enabled {
-            return VoiceBargeInTranscriptOutcome::BargeInDisabled;
-        }
-
-        if let Some(outcome) = self.apply_dispatcher_command(channel_id, transcript).await {
-            return outcome;
-        }
-
-        // #2250: in-flight foreground Codex/Claude calls are also
-        // cancellable "active work" — do not bail with NoActiveTurn if the
-        // only active work is a foreground call, otherwise barge-in cannot
-        // reach the registered cancel token.
-        let has_inflight_foreground = self
-            .channels
-            .inflight_foreground_cancels
-            .get(&channel_id.get())
-            .is_some_and(|entry| !entry.value().is_empty());
-        let cancel_channel = self
-            .active_barge_in_mailbox_channel(shared, channel_id)
-            .await;
-        if cancel_channel.is_none() && !has_inflight_foreground {
-            return VoiceBargeInTranscriptOutcome::NoActiveTurn;
-        }
-
-        let buffer = self.buffer_for_channel(channel_id);
-        let decision = buffer
-            .lock()
-            .await
-            .verify_processing_barge_in_after_stt(transcript);
-        match decision {
-            ProcessingBargeInDecision::AbortAgent => {
-                self.channels.barged_in(channel_id);
-                // #2250: also cancel any in-flight foreground/voice Codex
-                // call so its child process is killed mid-flight, not just
-                // the background turn.
-                let inflight_cancelled = self
-                    .cancel_inflight_foreground_calls(channel_id, "voice_barge_in_explicit_stop");
-                let _ = inflight_cancelled;
-                let cancel_channel = cancel_channel.unwrap_or(channel_id);
-                let result = super::mailbox_cancel_active_turn_with_reason(
-                    shared,
-                    cancel_channel,
-                    "voice_barge_in_explicit_stop",
-                )
-                .await;
-                // F22 (#2046): 사후 분석 라벨 강화. transcript 글자 수, 현재
-                // sensitivity, 활성 progress playback 보유 여부.
-                let sensitivity = self.current_sensitivity();
-                let playback_active = self.channels.playbacks.contains_key(&channel_id.get());
-                tracing::info!(
-                    channel_id = channel_id.get(),
-                    cancel_channel_id = cancel_channel.get(),
-                    cancelled = result.token.is_some(),
-                    already_stopping = result.already_stopping,
-                    transcript_chars = transcript.chars().count(),
-                    sensitivity = ?sensitivity,
-                    playback_active,
-                    "voice explicit-stop barge-in processed"
-                );
-                VoiceBargeInTranscriptOutcome::ExplicitStop {
-                    cancelled: result.token.is_some(),
-                    already_stopping: result.already_stopping,
-                    cancel_channel_id: cancel_channel.get(),
-                }
-            }
-            ProcessingBargeInDecision::DeferPrompt(prompt) => {
-                tracing::info!(
-                    channel_id = channel_id.get(),
-                    "voice processing barge-in deferred for next turn"
-                );
-                VoiceBargeInTranscriptOutcome::Deferred(prompt)
-            }
-            ProcessingBargeInDecision::IgnoreNoise => VoiceBargeInTranscriptOutcome::IgnoredNoise,
-        }
-    }
-
-    pub(in crate::services::discord) async fn try_handle_voice_transcript_announcement(
-        self: &Arc<Self>,
-        shared: &Arc<SharedData>,
-        source_channel_id: ChannelId,
-        announcement: &crate::voice::prompt::VoiceTranscriptAnnouncement,
-    ) -> bool {
-        if !self.enabled {
-            return false;
-        }
-        let Some(target_channel_id) = self
-            .resolve_voice_background_channel_for_source(source_channel_id)
-            .await
-        else {
-            tracing::warn!(
-                source_channel_id = source_channel_id.get(),
-                utterance_id = %announcement.utterance_id,
-                "voice foreground handoff skipped because no background channel is mapped"
-            );
-            return false;
-        };
-
-        let started = Instant::now();
-        let language = announcement.language.clone();
-        let foreground = self
-            .resolve_effective_foreground_config(source_channel_id, target_channel_id)
-            .await;
-        // #3906 (P1): redundant foreground-start chime removed — superseded by the
-        // deterministic Phase-1 intake chime in process_completed_utterance.
-        let cancel_token = Arc::new(crate::services::provider::CancelToken::new());
-        // #3911 + #2335 (c): the guard OWNS the registration. It registers the
-        // token BEFORE the generate `.await` below and unregisters on drop, so
-        // an abort mid-`.await` (shutdown / supervisor abort) cannot leak the
-        // token in `inflight_foreground_cancels` (a leak would keep
-        // `has_inflight_foreground` permanently true and misroute the next
-        // fresh utterance as a barge-in). Keeping the guard alive through every
-        // suppressible side effect below (synth, play, background dispatch)
-        // still lets a late cancel flip this token and suppress the stale
-        // ack/handoff; we re-check the cancel flag at each awaited boundary.
-        let _inflight_guard =
-            InflightForegroundCancelGuard::register(self, source_channel_id, cancel_token.clone());
-        let decision = self
-            .generate_foreground_ack_text_for_runtime(
-                &announcement.transcript,
-                &language,
-                &foreground,
-                cancel_token.clone(),
-            )
-            .await
-            .unwrap_or(VoiceForegroundDecision::Silence);
-        let foreground_latency_ms = started.elapsed().as_millis().min(u64::MAX as u128) as u64;
-
-        let record_cancel_suppressed = |label: &'static str| {
-            let mut event = voice_flight_event_from_announcement(
-                VoiceFlightRoute::ExplicitStop,
-                source_channel_id,
-                Some(target_channel_id),
-                announcement,
-            );
-            attach_foreground_flight_metadata(
-                &mut event,
-                &foreground,
-                foreground_latency_ms,
-                "cancelled",
-            );
-            event.cancel_source = cancel_token
-                .cancel_source()
-                .or_else(|| Some(label.to_string()));
-            event.cancel_channel_id = Some(target_channel_id.get());
-            event.cancelled = Some(true);
-            event.reason = Some(label.to_string());
-            record_voice_flight_event(event);
-        };
-
-        let log_cancel_suppressed = |label: &'static str| {
-            tracing::info!(
-                source_channel_id = source_channel_id.get(),
-                target_channel_id = target_channel_id.get(),
-                utterance_id = %announcement.utterance_id,
-                cancel_source = ?cancel_token.cancel_source(),
-                cancel_source_kind = ?cancel_token.cancel_source_kind(),
-                stage = label,
-                "voice foreground side effect suppressed because cancel won the race (#2335)"
-            );
-        };
-
-        if foreground_decision::ack_cancel_suppresses_fallback(&cancel_token) {
-            record_cancel_suppressed("post_generation");
-            log_cancel_suppressed("post_generation");
-            return true;
-        }
-
-        match decision {
-            VoiceForegroundDecision::Silence => {
-                let mut event = voice_flight_event_from_announcement(
-                    VoiceFlightRoute::ForegroundSilence,
-                    source_channel_id,
-                    Some(target_channel_id),
-                    announcement,
-                );
-                attach_foreground_flight_metadata(
-                    &mut event,
-                    &foreground,
-                    foreground_latency_ms,
-                    "silence",
-                );
-                record_voice_flight_event(event);
-                tracing::info!(
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    utterance_id = %announcement.utterance_id,
-                    elapsed_ms = started.elapsed().as_millis(),
-                    foreground_provider = %foreground.provider,
-                    foreground_model = %foreground.model,
-                    "voice foreground chose silence"
-                );
-            }
-            VoiceForegroundDecision::Speak(spoken) => {
-                if foreground_decision::ack_cancel_suppresses_fallback(&cancel_token) {
-                    record_cancel_suppressed("pre_speak_synth");
-                    log_cancel_suppressed("pre_speak_synth");
-                    return true;
-                }
-                if let Some(path) = self
-                    .synthesize_acknowledgement(&spoken, source_channel_id)
-                    .await
-                {
-                    if foreground_decision::ack_cancel_suppresses_fallback(&cancel_token) {
-                        record_cancel_suppressed("post_speak_synth");
-                        log_cancel_suppressed("post_speak_synth");
-                        return true;
-                    }
-                    self.play_acknowledgement(shared, source_channel_id, path)
-                        .await;
-                }
-                let mut event = voice_flight_event_from_announcement(
-                    VoiceFlightRoute::ForegroundSpeak,
-                    source_channel_id,
-                    Some(target_channel_id),
-                    announcement,
-                );
-                attach_foreground_flight_metadata(
-                    &mut event,
-                    &foreground,
-                    foreground_latency_ms,
-                    "speak",
-                );
-                event.tts_chars = Some(spoken.chars().count());
-                record_voice_flight_event(event);
-                tracing::info!(
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    utterance_id = %announcement.utterance_id,
-                    elapsed_ms = started.elapsed().as_millis(),
-                    foreground_provider = %foreground.provider,
-                    foreground_model = %foreground.model,
-                    "voice foreground spoken response queued"
-                );
-            }
-            VoiceForegroundDecision::HandoffBackground(summary) => {
-                if cancel_token.cancelled.load(Ordering::Relaxed) {
-                    record_cancel_suppressed("pre_background_handoff");
-                    log_cancel_suppressed("pre_background_handoff");
-                    return true;
-                }
-                match self
-                    .dispatch_voice_background_handoff(
-                        shared,
-                        source_channel_id,
-                        target_channel_id,
-                        announcement,
-                        &summary,
-                    )
-                    .await
-                {
-                    Ok(handoff_outcome) => {
-                        let VoiceBackgroundHandoffOutcome {
-                            turn_id,
-                            handoff_message_id,
-                            correlation_id,
-                        } = handoff_outcome;
-                        let tombstone = handoff_message_id.and_then(|id| {
-                            crate::voice::cancel_tombstone::global_store().lookup(id)
-                        });
-                        if let Some(observation) =
-                            observe_voice_handoff_cancel(&cancel_token, tombstone)
-                        {
-                            record_and_cancel_voice_handoff_if_observed(
-                                shared,
-                                source_channel_id,
-                                target_channel_id,
-                                &turn_id,
-                                handoff_message_id,
-                                observation.clone(),
-                            )
-                            .await;
-                            let mut event = voice_flight_event_from_announcement(
-                                VoiceFlightRoute::ExplicitStop,
-                                source_channel_id,
-                                Some(target_channel_id),
-                                announcement,
-                            );
-                            attach_foreground_flight_metadata(
-                                &mut event,
-                                &foreground,
-                                foreground_latency_ms,
-                                "handoff_cancelled",
-                            );
-                            event.handoff_correlation_id = Some(correlation_id.clone());
-                            event.handoff_message_id = handoff_message_id.map(|id| id.get());
-                            event.background_turn_id = Some(turn_id.clone());
-                            event.cancel_source = Some(observation.cancel_reason);
-                            event.cancel_channel_id = Some(target_channel_id.get());
-                            event.cancelled = Some(true);
-                            event.reason = Some("post_background_handoff_started".to_string());
-                            record_voice_flight_event(event);
-                            log_cancel_suppressed("post_background_handoff_started");
-                            return true;
-                        }
-                        let ack = voice_background_handoff_ack(&language);
-                        let ack_path = self
-                            .synthesize_acknowledgement(ack, source_channel_id)
-                            .await;
-                        // #2403: re-use the same cancel/tombstone handling
-                        // after synthesis. A stop arriving during TTS must
-                        // suppress the spoken ack AND cancel the just-started
-                        // target turn, not merely return before playback.
-                        let tombstone_after_synth = handoff_message_id.and_then(|id| {
-                            crate::voice::cancel_tombstone::global_store().lookup(id)
-                        });
-                        if let Some(observation) =
-                            observe_voice_handoff_cancel(&cancel_token, tombstone_after_synth)
-                        {
-                            record_and_cancel_voice_handoff_if_observed(
-                                shared,
-                                source_channel_id,
-                                target_channel_id,
-                                &turn_id,
-                                handoff_message_id,
-                                observation.clone(),
-                            )
-                            .await;
-                            let mut event = voice_flight_event_from_announcement(
-                                VoiceFlightRoute::ExplicitStop,
-                                source_channel_id,
-                                Some(target_channel_id),
-                                announcement,
-                            );
-                            attach_foreground_flight_metadata(
-                                &mut event,
-                                &foreground,
-                                foreground_latency_ms,
-                                "handoff_cancelled",
-                            );
-                            event.handoff_correlation_id = Some(correlation_id.clone());
-                            event.handoff_message_id = handoff_message_id.map(|id| id.get());
-                            event.background_turn_id = Some(turn_id.clone());
-                            event.cancel_source = Some(observation.cancel_reason);
-                            event.cancel_channel_id = Some(target_channel_id.get());
-                            event.cancelled = Some(true);
-                            event.reason = Some("post_background_handoff_play".to_string());
-                            record_voice_flight_event(event);
-                            log_cancel_suppressed("post_background_handoff_play");
-                            return true;
-                        }
-                        if let Some(path) = ack_path {
-                            self.play_acknowledgement(shared, source_channel_id, path)
-                                .await;
-                        }
-                        let mut event = voice_flight_event_from_announcement(
-                            VoiceFlightRoute::BackgroundHandoff,
-                            source_channel_id,
-                            Some(target_channel_id),
-                            announcement,
-                        );
-                        attach_foreground_flight_metadata(
-                            &mut event,
-                            &foreground,
-                            foreground_latency_ms,
-                            "handoff_background",
-                        );
-                        event.handoff_correlation_id = Some(correlation_id.clone());
-                        event.handoff_message_id = handoff_message_id.map(|id| id.get());
-                        event.background_turn_id = Some(turn_id.clone());
-                        event.tts_chars = Some(ack.chars().count());
-                        record_voice_flight_event(event);
-                        tracing::info!(
-                            source_channel_id = source_channel_id.get(),
-                            target_channel_id = target_channel_id.get(),
-                            utterance_id = %announcement.utterance_id,
-                            turn_id = %turn_id,
-                            elapsed_ms = started.elapsed().as_millis(),
-                            foreground_provider = %foreground.provider,
-                            foreground_model = %foreground.model,
-                            "voice foreground handed request to background"
-                        );
-                    }
-                    Err(error) => {
-                        let mut event = voice_flight_event_from_announcement(
-                            VoiceFlightRoute::BackgroundHandoff,
-                            source_channel_id,
-                            Some(target_channel_id),
-                            announcement,
-                        );
-                        attach_foreground_flight_metadata(
-                            &mut event,
-                            &foreground,
-                            foreground_latency_ms,
-                            "handoff_background",
-                        );
-                        event.reason = Some(format!("handoff_failed:{error}"));
-                        record_voice_flight_event(event);
-                        tracing::warn!(
-                            error = %error,
-                            source_channel_id = source_channel_id.get(),
-                            target_channel_id = target_channel_id.get(),
-                            utterance_id = %announcement.utterance_id,
-                            "voice foreground background handoff failed"
-                        );
-                    }
-                }
-            }
-        }
-        true
-    }
-
-    async fn dispatch_voice_background_handoff(
-        &self,
-        shared: &Arc<SharedData>,
-        source_channel_id: ChannelId,
-        target_channel_id: ChannelId,
-        announcement: &crate::voice::prompt::VoiceTranscriptAnnouncement,
-        summary: &str,
-    ) -> Result<VoiceBackgroundHandoffOutcome, String> {
-        let driver = select_voice_background_driver();
-        let guild_id = self.voice_turn_guild_id(source_channel_id, target_channel_id);
-        let prompt = build_voice_background_handoff_prompt(
-            &announcement.transcript,
-            summary,
-            &announcement.language,
-        );
-        let correlation_id = crate::voice::prompt::new_voice_background_handoff_correlation_id();
-        let prompt =
-            crate::voice::prompt::append_voice_background_handoff_marker(&prompt, &correlation_id);
-        let generation = default_voice_announce_generation() + 1;
-        let agent_id = self
-            .channels
-            .active_voice_routes
-            .get(&source_channel_id.get())
-            .map(|entry| entry.agent_id.clone());
-        let meta = crate::voice::announce_meta::VoiceBackgroundHandoffMeta {
-            voice_channel_id: source_channel_id.get(),
-            background_channel_id: target_channel_id.get(),
-            agent_id,
-            local_only_fallback: false,
-        };
-        let store = crate::voice::announce_meta::global_store();
-        store.reserve_handoff(&correlation_id, meta.clone());
-        let mut durable_reserved = false;
-        if let Some(pool) = shared.pg_pool.as_ref() {
-            match crate::voice::announce_meta::persist_handoff_reservation_durable(
-                pool,
-                &correlation_id,
-                &meta,
-            )
-            .await
-            {
-                Ok(()) => {
-                    durable_reserved = true;
-                }
-                Err(error) => {
-                    tracing::warn!(
-                        error = %error,
-                        correlation_id = %correlation_id,
-                        source_channel_id = source_channel_id.get(),
-                        target_channel_id = target_channel_id.get(),
-                        utterance_id = %announcement.utterance_id,
-                        "voice background handoff durable reservation failed before publish; refusing to publish"
-                    );
-                    store.cancel_handoff_reservation(&correlation_id);
-                    return Err(format!(
-                        "voice background handoff durable reservation failed before publish: {error}"
-                    ));
-                }
-            }
-        } else {
-            tracing::debug!(
-                correlation_id = %correlation_id,
-                "voice background handoff durable reservation skipped — postgres pool unavailable"
-            );
-        }
-
-        let link_generation = generation.min(i32::MAX as u64) as i32;
-        let mut voice_turn_link_inserted = false;
-        if let (Some(pool), Some(guild_id)) = (shared.pg_pool.as_ref(), guild_id) {
-            let link = crate::voice::turn_link::VoiceTurnLinkInsert {
-                guild_id: guild_id.get(),
-                voice_channel_id: source_channel_id.get(),
-                background_channel_id: target_channel_id.get(),
-                utterance_id: announcement.utterance_id.clone(),
-                generation: link_generation,
-                announce_message_id: None,
-                dispatch_id: None,
-                turn_id: None,
-            };
-            match crate::voice::turn_link::upsert_active_voice_turn_link_pg(pool, &link).await {
-                Ok(Some(_)) => {
-                    voice_turn_link_inserted = true;
-                }
-                Ok(None) => {}
-                Err(error) => {
-                    tracing::warn!(
-                        error = %error,
-                        source_channel_id = source_channel_id.get(),
-                        target_channel_id = target_channel_id.get(),
-                        utterance_id = %announcement.utterance_id,
-                        "voice background handoff voice_turn_link pre-publish insert failed; terminal TTS will fall back to announce metadata"
-                    );
-                }
-            }
-        }
-
-        let start_result = {
-            #[cfg(test)]
-            {
-                if let Some(result) = self.take_test_background_handoff_outcome(
-                    driver.kind(),
-                    source_channel_id,
-                    target_channel_id,
-                    announcement,
-                    summary,
-                    &prompt,
-                ) {
-                    result
-                } else {
-                    driver
-                        .start(VoiceBackgroundStartRequest {
-                            guild_id,
-                            voice_channel_id: source_channel_id,
-                            channel_id: target_channel_id,
-                            shared,
-                            utterance_id: &announcement.utterance_id,
-                            generation,
-                            message_content: &prompt,
-                        })
-                        .await
-                }
-            }
-            #[cfg(not(test))]
-            {
-                driver
-                    .start(VoiceBackgroundStartRequest {
-                        guild_id,
-                        voice_channel_id: source_channel_id,
-                        channel_id: target_channel_id,
-                        shared,
-                        utterance_id: &announcement.utterance_id,
-                        generation,
-                        message_content: &prompt,
-                    })
-                    .await
-            }
-        };
-        let outcome = match start_result {
-            Ok(outcome) => outcome,
-            Err(error) => {
-                store.cancel_handoff_reservation(&correlation_id);
-                if voice_turn_link_inserted {
-                    if let (Some(pool), Some(guild_id)) = (shared.pg_pool.as_ref(), guild_id) {
-                        if let Err(mark_error) =
-                            crate::voice::turn_link::mark_terminal_voice_turn_link_pg(
-                                pool,
-                                guild_id.get(),
-                                source_channel_id.get(),
-                                &announcement.utterance_id,
-                                link_generation,
-                            )
-                            .await
-                        {
-                            tracing::warn!(
-                                error = %mark_error,
-                                correlation_id = %correlation_id,
-                                "voice_turn_link cleanup failed after background handoff publish error"
-                            );
-                        }
-                    }
-                }
-                if durable_reserved {
-                    if let Some(pool) = shared.pg_pool.as_ref() {
-                        if let Err(cancel_error) =
-                            crate::voice::announce_meta::cancel_handoff_reservation_durable(
-                                pool,
-                                &correlation_id,
-                            )
-                            .await
-                        {
-                            tracing::warn!(
-                                error = %cancel_error,
-                                correlation_id = %correlation_id,
-                                "voice background handoff durable reservation cleanup failed after publish error"
-                            );
-                        }
-                    }
-                }
-                return Err(error);
-            }
-        };
-
-        if let (Some(pool), Some(guild_id)) = (shared.pg_pool.as_ref(), guild_id) {
-            let patch = crate::voice::turn_link::VoiceTurnLinkIdentityPatch {
-                guild_id: guild_id.get(),
-                voice_channel_id: source_channel_id.get(),
-                utterance_id: announcement.utterance_id.clone(),
-                generation: link_generation,
-                announce_message_id: outcome.message_id.map(|id| id.get()),
-                dispatch_id: None,
-                turn_id: Some(outcome.turn_id.clone()),
-            };
-            if let Err(error) =
-                crate::voice::turn_link::attach_voice_turn_link_ids_pg(pool, &patch).await
-            {
-                tracing::warn!(
-                    error = %error,
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    utterance_id = %announcement.utterance_id,
-                    announce_message_id = outcome.message_id.map(|id| id.get()),
-                    turn_id = %outcome.turn_id,
-                    "voice background handoff voice_turn_link identity attach failed; terminal TTS will fall back to announce metadata"
-                );
-            }
-        }
-
-        // #2236: stamp a typed marker keyed by the posted message id so the
-        // turn bridge can route the background turn's spoken summary back to
-        // the foreground voice channel WITHOUT relying on user-authored prompt
-        // text. #2392 moves the stamp to a pre-publish reservation: if an
-        // immediate completion beats this bind, terminal delivery claims the
-        // reservation by the correlation marker and this late bind becomes a
-        // no-op instead of resurrecting the handoff.
-        if let Some(message_id) = outcome.message_id {
-            if !store.bind_handoff_message_id(&correlation_id, message_id) {
-                tracing::info!(
-                    correlation_id = %correlation_id,
-                    message_id = message_id.get(),
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    utterance_id = %announcement.utterance_id,
-                    "voice background handoff local reservation was already consumed before message-id bind"
-                );
-            }
-            if durable_reserved {
-                if let Some(pool) = shared.pg_pool.as_ref() {
-                    match crate::voice::announce_meta::bind_handoff_durable_message_id(
-                        pool,
-                        &correlation_id,
-                        message_id,
-                    )
-                    .await
-                    {
-                        Ok(true) => {}
-                        Ok(false) => {
-                            tracing::info!(
-                                correlation_id = %correlation_id,
-                                message_id = message_id.get(),
-                                source_channel_id = source_channel_id.get(),
-                                target_channel_id = target_channel_id.get(),
-                                utterance_id = %announcement.utterance_id,
-                                "voice background handoff durable reservation was already consumed before message-id bind"
-                            );
-                        }
-                        Err(error) => {
-                            tracing::warn!(
-                                error = %error,
-                                correlation_id = %correlation_id,
-                                message_id = message_id.get(),
-                                source_channel_id = source_channel_id.get(),
-                                target_channel_id = target_channel_id.get(),
-                                utterance_id = %announcement.utterance_id,
-                                "voice background handoff durable reservation bind failed"
-                            );
-                        }
-                    }
-                }
-            }
-        } else {
-            store.cancel_handoff_reservation(&correlation_id);
-            if durable_reserved {
-                if let Some(pool) = shared.pg_pool.as_ref() {
-                    if let Err(error) =
-                        crate::voice::announce_meta::cancel_handoff_reservation_durable(
-                            pool,
-                            &correlation_id,
-                        )
-                        .await
-                    {
-                        tracing::warn!(
-                            error = %error,
-                            correlation_id = %correlation_id,
-                            "voice background handoff durable reservation cleanup failed after missing message id"
-                        );
-                    }
-                }
-            }
-            tracing::warn!(
-                source_channel_id = source_channel_id.get(),
-                target_channel_id = target_channel_id.get(),
-                utterance_id = %announcement.utterance_id,
-                "voice background handoff dispatch returned no message_id; spoken summary routing will fall back to legacy prefix detection"
-            );
-        }
-        Ok(VoiceBackgroundHandoffOutcome {
-            turn_id: outcome.turn_id,
-            handoff_message_id: outcome.message_id,
-            correlation_id,
-        })
-    }
-
-    async fn start_voice_turn(
-        self: &Arc<Self>,
-        shared: &Arc<SharedData>,
-        source_channel_id: ChannelId,
-        target_channel_id: ChannelId,
-        utterance: &CompletedUtterance,
-        transcript: &str,
-        flight_context: &VoiceFlightUtteranceContext,
-    ) -> VoiceBargeInTranscriptOutcome {
-        let verbose_progress = self.verbose_progress_enabled();
-        let language = self.spoken_result_language().await;
-        let foreground = self
-            .resolve_effective_foreground_config(source_channel_id, target_channel_id)
-            .await;
-        let driver = select_voice_background_driver();
-        let guild_id = self.voice_turn_guild_id(source_channel_id, target_channel_id);
-        if guild_id.is_none() {
-            tracing::warn!(
-                source_channel_id = source_channel_id.get(),
-                target_channel_id = target_channel_id.get(),
-                utterance_id = %utterance.utterance_id,
-                "voice transcript announcement has no registered voice guild; sending without delivery id"
-            );
-        }
-        let announcement = crate::voice::prompt::build_voice_transcript_announcement(
-            transcript,
-            utterance.user_id,
-            &utterance.utterance_id,
-            &language,
-            verbose_progress,
-            &utterance.started_at,
-            &utterance.completed_at,
-            utterance.samples_written,
-        );
-        let announcement_meta = crate::voice::prompt::voice_transcript_announcement_meta(
-            transcript,
-            utterance.user_id,
-            &utterance.utterance_id,
-            &language,
-            verbose_progress,
-            &utterance.started_at,
-            &utterance.completed_at,
-            utterance.samples_written,
-            Some(
-                utterance
-                    .control_channel_id
-                    .unwrap_or(source_channel_id.get()),
-            ),
-            Some(flight_context.stt_mode.as_str()),
-            Some(flight_context.stt_latency_ms),
-        );
-        let generation = default_voice_announce_generation();
-        let voice_delivery_id = guild_id.map(|guild_id| {
-            voice_announce_delivery_id(
-                guild_id,
-                source_channel_id,
-                &utterance.utterance_id,
-                generation,
-            )
-        });
-        let durable_pending_key = voice_delivery_id
-            .as_ref()
-            .map(|delivery_id| {
-                crate::voice::announce_meta::durable_voice_announcement_pending_key(
-                    &delivery_id.correlation_id,
-                    &delivery_id.semantic_event_id,
-                )
-            })
-            .unwrap_or_else(|| {
-                crate::voice::announce_meta::durable_voice_announcement_pending_key(
-                    &format!(
-                        "voice:no-guild:{}:{}:{}",
-                        source_channel_id.get(),
-                        target_channel_id.get(),
-                        utterance.utterance_id
-                    ),
-                    &format!("announce:generation:{generation}"),
-                )
-            });
-        let announcement = crate::voice::prompt::append_voice_transcript_announcement_ref(
-            &announcement,
-            &durable_pending_key,
-        );
-        let mut durable_reserved = false;
-        if let Some(pool) = shared.pg_pool.as_ref() {
-            match crate::voice::announce_meta::persist_voice_announcement_reservation_durable(
-                pool,
-                &durable_pending_key,
-                source_channel_id,
-                &announcement,
-                &announcement_meta,
-            )
-            .await
-            {
-                Ok(true) => {
-                    durable_reserved = true;
-                }
-                Ok(false) => {
-                    tracing::warn!(
-                        source_channel_id = source_channel_id.get(),
-                        target_channel_id = target_channel_id.get(),
-                        utterance_id = %utterance.utterance_id,
-                        "voice transcript announcement durable reservation was already consumed; refusing to resurrect voice metadata"
-                    );
-                    return VoiceBargeInTranscriptOutcome::VoiceTurnStartFailed(
-                        "voice transcript announcement was already consumed".to_string(),
-                    );
-                }
-                Err(error) => {
-                    tracing::warn!(
-                        error = %error,
-                        source_channel_id = source_channel_id.get(),
-                        target_channel_id = target_channel_id.get(),
-                        utterance_id = %utterance.utterance_id,
-                        "voice transcript announcement durable reservation failed before publish"
-                    );
-                    return VoiceBargeInTranscriptOutcome::VoiceTurnStartFailed(format!(
-                        "voice transcript announcement durable reservation failed: {error}"
-                    ));
-                }
-            }
-        }
-        let start_result = {
-            #[cfg(test)]
-            {
-                if let Some(outcome) = self.take_test_turn_start_outcome(
-                    driver.kind(),
-                    source_channel_id,
-                    target_channel_id,
-                    &utterance.utterance_id,
-                    &announcement,
-                ) {
-                    outcome
-                } else {
-                    driver
-                        .start(VoiceBackgroundStartRequest {
-                            guild_id,
-                            voice_channel_id: source_channel_id,
-                            channel_id: source_channel_id,
-                            shared,
-                            utterance_id: &utterance.utterance_id,
-                            generation,
-                            message_content: &announcement,
-                        })
-                        .await
-                }
-            }
-            #[cfg(not(test))]
-            {
-                driver
-                    .start(VoiceBackgroundStartRequest {
-                        guild_id,
-                        voice_channel_id: source_channel_id,
-                        channel_id: source_channel_id,
-                        shared,
-                        utterance_id: &utterance.utterance_id,
-                        generation,
-                        message_content: &announcement,
-                    })
-                    .await
-            }
-        };
-        match start_result {
-            Ok(outcome) => {
-                if let Some(message_id) = outcome.message_id {
-                    let mut cache_local_metadata = !durable_reserved;
-                    if durable_reserved {
-                        if let Some(pool) = shared.pg_pool.as_ref() {
-                            match crate::voice::announce_meta::bind_voice_announcement_durable_message_id(
-                                pool,
-                                &durable_pending_key,
-                                message_id,
-                            )
-                            .await
-                            {
-                                Ok(true) => {
-                                    cache_local_metadata = true;
-                                }
-                                Ok(false) => {
-                                    tracing::info!(
-                                        message_id = message_id.get(),
-                                        source_channel_id = source_channel_id.get(),
-                                        target_channel_id = target_channel_id.get(),
-                                        utterance_id = %utterance.utterance_id,
-                                        "voice transcript announcement durable reservation was already consumed or bound elsewhere; skipping local metadata cache"
-                                    );
-                                }
-                                Err(error) => {
-                                    tracing::warn!(
-                                        error = %error,
-                                        message_id = message_id.get(),
-                                        source_channel_id = source_channel_id.get(),
-                                        target_channel_id = target_channel_id.get(),
-                                        utterance_id = %utterance.utterance_id,
-                                        "voice transcript announcement durable reservation bind failed; skipping local metadata cache so workers must use the pending ref"
-                                    );
-                                }
-                            }
-                        }
-                    }
-                    if cache_local_metadata {
-                        crate::voice::announce_meta::global_store()
-                            .insert(message_id, announcement_meta);
-                    }
-                }
-                tracing::info!(
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    user_id = utterance.user_id,
-                    utterance_id = %utterance.utterance_id,
-                    turn_id = %outcome.turn_id,
-                    background_driver = %outcome.driver_kind.as_str(),
-                    foreground_provider = %foreground.provider,
-                    foreground_model = %foreground.model,
-                    foreground_max_chars = foreground.max_chars,
-                    "voice transcript announcement posted to voice channel as canonical foreground trigger"
-                );
-                let mut event = flight_context.event(VoiceFlightRoute::Queued);
-                event.background_channel_id = Some(target_channel_id.get());
-                event.turn_id = Some(outcome.turn_id.clone());
-                event.foreground_provider = Some(foreground.provider.clone());
-                event.foreground_model = Some(foreground.model.clone());
-                event.foreground_decision = Some("queued_foreground_trigger".to_string());
-                if let Some(route) = self
-                    .channels
-                    .active_voice_routes
-                    .get(&source_channel_id.get())
-                {
-                    event.agent_id = Some(route.agent_id.clone());
-                }
-                record_voice_flight_event(event);
-                return VoiceBargeInTranscriptOutcome::VoiceTurnStarted {
-                    turn_id: outcome.turn_id,
-                };
-            }
-            Err(error) => {
-                if durable_reserved {
-                    if let Some(pool) = shared.pg_pool.as_ref() {
-                        if let Err(cancel_error) =
-                            crate::voice::announce_meta::cancel_voice_announcement_reservation_durable(
-                                pool,
-                                &durable_pending_key,
-                            )
-                            .await
-                        {
-                            tracing::warn!(
-                                error = %cancel_error,
-                                source_channel_id = source_channel_id.get(),
-                                target_channel_id = target_channel_id.get(),
-                                utterance_id = %utterance.utterance_id,
-                                "voice transcript announcement durable reservation cleanup failed after publish error"
-                            );
-                        }
-                    }
-                }
-                tracing::warn!(
-                    error = %error,
-                    source_channel_id = source_channel_id.get(),
-                    target_channel_id = target_channel_id.get(),
-                    user_id = utterance.user_id,
-                    utterance_id = %utterance.utterance_id,
-                    "voice transcript announcement failed; refusing direct voice turn fallback"
-                );
-                let mut event = flight_context.event(VoiceFlightRoute::Queued);
-                event.background_channel_id = Some(target_channel_id.get());
-                event.foreground_provider = Some(foreground.provider.clone());
-                event.foreground_model = Some(foreground.model.clone());
-                event.foreground_decision = Some("queue_failed".to_string());
-                event.reason = Some(format!("voice_turn_start_failed:{error}"));
-                record_voice_flight_event(event);
-                return VoiceBargeInTranscriptOutcome::VoiceTurnStartFailed(format!(
-                    "voice transcript announcement failed: {error}"
-                ));
-            }
-        }
-    }
-
-    /// F6 (#2046): `Config` 핫캐시. TTL 안이면 캐시된 `Arc<Config>` 반환,
-    /// 만료 시 spawn_blocking 으로 `load_graceful` 을 1회 호출해 갱신한다.
-    /// 매 utterance 마다 동기 std::fs read + serde_yaml 파싱이 발생하던 hot path 를
-    /// 5초 TTL 로 묶어 부하를 줄이고 async executor 블록도 회피한다.
-    async fn cached_config(&self) -> Arc<crate::config::Config> {
-        let now = Instant::now();
-        if let Some(cached) = self.config_cache.lookup_within_ttl(now) {
-            return cached;
-        }
-        let fresh = tokio::task::spawn_blocking(crate::config::load_graceful)
-            .await
-            .unwrap_or_else(|_| crate::config::Config::default());
-        let arc = Arc::new(fresh);
-        self.config_cache.store(Instant::now(), arc.clone());
-        arc
-    }
-
-    pub(in crate::services::discord) async fn process_completed_utterance(
-        self: &Arc<Self>,
-        shared: &Arc<SharedData>,
-        provider: &ProviderKind,
-        channel_id: ChannelId,
-        utterance: &CompletedUtterance,
-    ) -> VoiceBargeInTranscriptOutcome {
-        if !self.enabled {
-            return VoiceBargeInTranscriptOutcome::Disabled;
-        }
-
-        let transcribed = match self
-            .transcribe_completed_utterance(channel_id, utterance)
-            .await
-        {
-            Some(transcript) => transcript,
-            None => return VoiceBargeInTranscriptOutcome::TranscriptUnavailable,
-        };
-
-        let transcript = transcribed.text.trim();
-        let config_snapshot = crate::config::load_graceful();
-        let source_channel_id = effective_voice_source_channel(&config_snapshot, channel_id);
-        let flight_context = VoiceFlightUtteranceContext::from_utterance(
-            source_channel_id,
-            utterance,
-            transcript,
-            &transcribed,
-        );
-        if transcript.is_empty() {
-            let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-            event.reason = Some("empty_transcript".to_string());
-            record_voice_flight_event(event);
-            return VoiceBargeInTranscriptOutcome::EmptyTranscript;
-        }
-
-        let source_is_agent_voice = config_snapshot
-            .agents
-            .iter()
-            .any(|agent| agent_voice_matches_channel(agent, source_channel_id));
-        let source_is_lobby = super::settings::resolve_role_binding(source_channel_id, None)
-            .is_none()
-            && voice_lobby_accepts_source_channel(&config_snapshot.voice, source_channel_id);
-        let transcript = if source_is_agent_voice || source_is_lobby {
-            transcript.to_string()
-        } else {
-            match self.runtime_wake_word_decision(transcript).await {
-                WakeWordDecision::NotRequired(transcript) => transcript,
-                WakeWordDecision::Matched(matched) => matched.remaining,
-                WakeWordDecision::Missing => {
-                    let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-                    event.reason = Some("wake_word_required".to_string());
-                    record_voice_flight_event(event);
-                    return VoiceBargeInTranscriptOutcome::WakeWordRequired;
-                }
-            }
-        };
-        let transcript = transcript.trim();
-        if transcript.is_empty() {
-            let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-            event.reason = Some("empty_after_wake_word".to_string());
-            record_voice_flight_event(event);
-            return VoiceBargeInTranscriptOutcome::EmptyTranscript;
-        }
-
-        // #2250: also treat in-flight foreground Codex/Claude calls as
-        // active work for barge-in purposes. Otherwise a barge-in arriving
-        // while we are still generating the ack / channel-text / summary
-        // would bypass `handle_processing_transcript` and never cancel the
-        // spawned child.
-        let has_inflight_foreground = self
-            .channels
-            .inflight_foreground_cancels
-            .get(&source_channel_id.get())
-            .is_some_and(|entry| !entry.value().is_empty());
-        if self
-            .active_barge_in_mailbox_channel(shared, source_channel_id)
-            .await
-            .is_some()
-            || has_inflight_foreground
-        {
-            let outcome = self
-                .handle_processing_transcript(shared, provider, source_channel_id, transcript)
-                .await;
-            let mut event = match &outcome {
-                VoiceBargeInTranscriptOutcome::ExplicitStop {
-                    cancelled,
-                    already_stopping,
-                    cancel_channel_id,
-                } => {
-                    let mut event = flight_context.event(VoiceFlightRoute::ExplicitStop);
-                    event.cancel_channel_id = Some(*cancel_channel_id);
-                    event.barge_in = Some(true);
-                    event.cancel_source = Some("voice_barge_in_explicit_stop".to_string());
-                    event.cancelled = Some(*cancelled);
-                    event.already_stopping = Some(*already_stopping);
-                    event
-                }
-                VoiceBargeInTranscriptOutcome::Deferred(_) => {
-                    let mut event = flight_context.event(VoiceFlightRoute::Deferred);
-                    event.barge_in = Some(true);
-                    event.reason = Some("processing_barge_in_defer".to_string());
-                    event
-                }
-                VoiceBargeInTranscriptOutcome::IgnoredNoise => {
-                    let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-                    event.barge_in = Some(true);
-                    event.reason = Some("processing_barge_in_noise".to_string());
-                    event
-                }
-                _ => {
-                    let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-                    event.reason = Some("processing_barge_in_no_route".to_string());
-                    event
-                }
-            };
-            if let Some(route) = self
-                .channels
-                .active_voice_routes
-                .get(&source_channel_id.get())
-            {
-                event.agent_id = Some(route.agent_id.clone());
-                event.background_channel_id = Some(route.channel_id.get());
-            }
-            record_voice_flight_event(event);
-            return outcome;
-        }
-
-        if let Some(outcome) = self
-            .apply_dispatcher_command(source_channel_id, transcript)
-            .await
-        {
-            return outcome;
-        }
-
-        let (target_channel_id, transcript) = match self
-            .resolve_voice_turn_target(shared, source_channel_id, transcript)
-            .await
-        {
-            VoiceTurnTargetResolution::Target {
-                channel_id,
-                transcript,
-            } => (channel_id, transcript),
-            VoiceTurnTargetResolution::NeedsAgent => {
-                let mut event = flight_context.event(VoiceFlightRoute::Deferred);
-                event.reason = Some("agent_routing_required".to_string());
-                record_voice_flight_event(event);
-                self.ask_for_agent(shared, channel_id).await;
-                return VoiceBargeInTranscriptOutcome::AgentRoutingRequired;
-            }
-            VoiceTurnTargetResolution::Ignored => {
-                let mut event = flight_context.event(VoiceFlightRoute::IgnoredNoise);
-                event.reason = Some("no_active_voice_route".to_string());
-                record_voice_flight_event(event);
-                return VoiceBargeInTranscriptOutcome::NoActiveTurn;
-            }
-        };
-
-        // #3906 (P1): deterministic intake ack — Phase-1 chime BEFORE
-        // start_voice_turn, so it survives every VoiceTurnStartFailed/#3905 drop.
-        self.play_processing_chime(shared, source_channel_id).await;
-        self.start_voice_turn(
-            shared,
-            source_channel_id,
-            target_channel_id,
-            utterance,
-            &transcript,
-            &flight_context,
-        )
-        .await
-    }
-
-    pub(in crate::services::discord) async fn drain_deferred_after_turn(
-        self: &Arc<Self>,
-        shared: &Arc<SharedData>,
-        provider: &ProviderKind,
-        channel_id: ChannelId,
-    ) -> bool {
-        if !self.barge_in_enabled {
-            return false;
-        }
-
-        let Some(drain) = self.take_deferred_prompt(channel_id).await else {
-            return false;
-        };
-
-        if let Some(acknowledgement) = drain.acknowledgement {
-            if let Some(path) = self
-                .synthesize_acknowledgement(&acknowledgement, channel_id)
-                .await
-            {
-                self.play_acknowledgement(shared, channel_id, path).await;
-            }
-        }
-
-        let message_id = MessageId::new(self.id_sequences.next_internal_message_id());
-        super::enqueue_internal_followup(
-            shared,
-            provider,
-            channel_id,
-            message_id,
-            drain.prompt,
-            "voice barge-in deferred prompt",
-        )
-        .await
-    }
-
-    async fn take_deferred_prompt(&self, channel_id: ChannelId) -> Option<DeferredBargeInDrain> {
-        let buffer = self
-            .channels
-            .deferred_buffers
-            .get(&channel_id.get())
-            .map(|entry| entry.value().clone())?;
-        let mut buffer = buffer.lock().await;
-        let ack = &self.acknowledgement;
-        let acknowledgement = buffer
-            .acknowledgement_before_drain(ack.enabled(), ack.text())
-            .map(ToOwned::to_owned);
-        let prompt = buffer.drain_prompt()?;
-        Some(DeferredBargeInDrain {
-            acknowledgement,
-            prompt,
-        })
-    }
-
-    fn buffer_for_channel(&self, channel_id: ChannelId) -> Arc<Mutex<DeferredBargeInBuffer>> {
-        self.channels
-            .deferred_buffers
-            .entry(channel_id.get())
-            .or_insert_with(|| Arc::new(Mutex::new(DeferredBargeInBuffer::new())))
-            .clone()
-    }
-}
-
-fn pcm_i16_to_le_bytes(samples: &[i16]) -> Vec<u8> {
-    let mut bytes = Vec::with_capacity(samples.len() * 2);
-    for sample in samples {
-        bytes.extend_from_slice(&sample.to_le_bytes());
-    }
-    bytes
-}
-
-fn discord_pcm_i16_stereo_48k_to_mono_f32_16k(samples: &[i16]) -> Vec<f32> {
-    samples
-        .chunks_exact(6)
-        .map(|chunk| {
-            let mut sum = 0.0f32;
-            for frame in 0..3 {
-                let left = chunk[frame * 2] as f32;
-                let right = chunk[frame * 2 + 1] as f32;
-                sum += (left + right) * 0.5;
-            }
-            (sum / 3.0) / i16::MAX as f32
-        })
-        .collect()
-}
-
-/// #2156: 일반 wav/segment 삭제. NotFound 는 무시, 그 외 에러는 debug 로그.
-async fn remove_file_quietly(path: &Path) {
-    match tokio::fs::remove_file(path).await {
-        Ok(()) => {}
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
-        Err(error) => tracing::debug!(
-            error = %error,
-            path = %path.display(),
-            "voice utterance cleanup could not remove file (#2156)"
-        ),
-    }
-}
-
-/// #2156: transcript sidecar 정리. 후보 다수 중 대부분은 존재하지 않으므로
-/// 모든 에러를 trace 로 낮춰 로그 노이즈를 줄인다.
-async fn remove_file_quietly_silent(path: &Path) {
-    if let Err(error) = tokio::fs::remove_file(path).await
-        && error.kind() != std::io::ErrorKind::NotFound
-    {
-        tracing::trace!(
-            error = %error,
-            path = %path.display(),
-            "voice transcript sidecar cleanup skipped (#2156)"
-        );
-    }
-}
-
-fn transcript_dirs_from_config(config: &VoiceConfig) -> Vec<PathBuf> {
-    vec![expand_tilde(&config.audio.transcripts_dir)]
-}
-
-fn expand_tilde(path: &Path) -> PathBuf {
-    let raw = path.to_string_lossy();
-    if raw == "~" {
-        return dirs::home_dir().unwrap_or_else(|| path.to_path_buf());
-    }
-    if let Some(rest) = raw.strip_prefix("~/")
-        && let Some(home) = dirs::home_dir()
-    {
-        return home.join(rest);
-    }
-    path.to_path_buf()
-}
-
-fn lock_monitor(
-    monitor: &std::sync::Mutex<LiveBargeInMonitor>,
-) -> std::sync::MutexGuard<'_, LiveBargeInMonitor> {
-    monitor
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner())
 }
 
 #[cfg(test)]
